+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 - experiences uses new rule
 - feature changes: you can read the feature; you can quickly remove bonus
 - damnage in live manager show min max and mean damage 
-
+- css improvements
+- more suggested features
+- you can pick a new set of suggested features from another type
 
 # 0.1.5
 - small css fix